requirements.txt
langgraph>=0.0.18
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.10
pydantic>=2.5.3
typing-extensions>=4.9.0
python-dotenv>=1.0.0 